general updates

+++ lab9_project/display_artillery.c
@@ -136,7 +136,7 @@ void display_artillery_update_B_counter_display(int16_t count) {
   prevOnesPlace = onesPlace;
 }
 
-void display_artillery_update_P_counter_display(uint16_t count) {
+void display_artillery_update_P_counter_display(int16_t count) {
   static uint8_t prevTensPlace = 0;
   static uint8_t prevOnesPlace = 0;
 
@@ -157,7 +157,7 @@ void display_artillery_update_P_counter_display(uint16_t count) {
   prevOnesPlace = onesPlace;
 }
 
-void display_artillery_update_W_counter_display(uint8_t count) {
+void display_artillery_update_W_counter_display(int16_t count) {
   static uint8_t prevTensPlace = 0;
   static uint8_t prevOnesPlace = 0;
