+++ lab9_project/display_artillery.h
@@ -10,8 +10,8 @@ void display_artillery_timer_display(uint8_t count);
 
 void display_artillery_update_B_counter_display(int16_t count);
 
-void display_artillery_update_P_counter_display(uint16_t count);
+void display_artillery_update_P_counter_display(int16_t count);
 
-void display_artillery_update_W_counter_display(uint8_t count);
+void display_artillery_update_W_counter_display(int16_t count);
 
 #endif /* DISPLAY_ARTILLERY */
